+++ mods/mod-swooper-maps/src/recipes/standard/stages/map-hydrology/artifacts/hydrology-lakes-engine-terrain-snapshot.artifact.ts
@@ -21,7 +21,7 @@ export const Schema = Type.Object(
     landMask: TypedArraySchemas.u8({
       description: "Engine-derived land mask after lake projection (1=land, 0=water).",
     }),
-    terrain: TypedArraySchemas.u8({
+    terrain: TypedArraySchemas.i32({
       description: "Engine-derived terrain type snapshot after lake projection.",
     }),
     elevation: TypedArraySchemas.i16({
@@ -55,7 +55,7 @@ function validateLocal(
   const candidate = value as Record<string, unknown>;
   const cellCount = artifactCellCount(context);
   appendArtifactTypedArrayIssues(issues, "landMask", candidate.landMask, Uint8Array, cellCount);
-  appendArtifactTypedArrayIssues(issues, "terrain", candidate.terrain, Uint8Array, cellCount);
+  appendArtifactTypedArrayIssues(issues, "terrain", candidate.terrain, Int32Array, cellCount);
   appendArtifactTypedArrayIssues(issues, "elevation", candidate.elevation, Int16Array, cellCount);
   return Object.freeze(issues);
 }
