+++ mods/mod-swooper-maps/src/recipes/standard/stages/map-morphology/artifacts/continent-validation-terrain-snapshot.artifact.ts
@@ -21,7 +21,7 @@ export const Schema = Type.Object(
     landMask: TypedArraySchemas.u8({
       description: "Engine-derived land mask at this map-morphology boundary.",
     }),
-    terrain: TypedArraySchemas.u8({
+    terrain: TypedArraySchemas.i32({
       description: "Engine-derived terrain type snapshot at this map-morphology boundary.",
     }),
     elevation: TypedArraySchemas.i16({
@@ -52,7 +52,7 @@ function validateLocal(
   const candidate = value as Record<string, unknown>;
   const cellCount = artifactCellCount(context);
   appendArtifactTypedArrayIssues(issues, "landMask", candidate.landMask, Uint8Array, cellCount);
-  appendArtifactTypedArrayIssues(issues, "terrain", candidate.terrain, Uint8Array, cellCount);
+  appendArtifactTypedArrayIssues(issues, "terrain", candidate.terrain, Int32Array, cellCount);
   appendArtifactTypedArrayIssues(issues, "elevation", candidate.elevation, Int16Array, cellCount);
   return Object.freeze(issues);
 }
